refactor(api): deprecate POST /schooling-registration-user-associations/auto
in favor of POST /sco-organization-learners/association/auto
