+++ api/tests/integration/application/organization-learner-user-associations/index_test.js
@@ -11,9 +11,6 @@ describe('Integration | Application | Route | organization-learner-user-associat
       .stub(organizationLearnerUserAssociationController, 'reconcileSupOrganizationLearner')
       .callsFake((request, h) => h.response('ok').code(204));
     sinon
-      .stub(organizationLearnerUserAssociationController, 'reconcileOrganizationLearnerAutomatically')
-      .callsFake((request, h) => h.response('ok').code(204));
-    sinon
       .stub(organizationLearnerUserAssociationController, 'generateUsername')
       .callsFake((request, h) => h.response('ok').code(200));
 
@@ -232,46 +229,6 @@ describe('Integration | Application | Route | organization-learner-user-associat
         // then
         expect(response.statusCode).to.equal(422);
       });
-    });
-  });
-
-  describe('POST /api/schooling-registration-user-associations/auto', function () {
-    const method = 'POST';
-    const url = '/api/schooling-registration-user-associations/auto';
-
-    it('should succeed', async function () {
-      // given
-      const payload = {
-        data: {
-          attributes: {
-            'campaign-code': 'RESTRICTD',
-          },
-        },
-      };
-
-      // when
-      const response = await httpTestServer.request(method, url, payload);
-
-      // then
-      expect(response.statusCode).to.equal(204);
-    });
-
-    it('should return an error when there is an invalid campaign code attribute in the payload', async function () {
-      // given
-      const INVALID_CAMPAIGNCODE = '';
-      const payload = {
-        data: {
-          attributes: {
-            'campaign-code': INVALID_CAMPAIGNCODE,
-          },
-        },
-      };
-
-      // when
-      const response = await httpTestServer.request(method, url, payload);
-
-      // then
-      expect(response.statusCode).to.equal(422);
     });
   });
