+++ api/tests/unit/application/sco-organization-learners/sco-organization-learner-controller_test.js
@@ -47,4 +47,52 @@ describe('Unit | Application | Controller | sco-organization-learner', function 
       expect(response.headers['Deprecation']).to.not.exist;
     });
   });
+
+  describe('#reconcileScoOrganizationLearnerAutomatically', function () {
+    const userId = 2;
+    const request = {
+      auth: { credentials: { userId } },
+      payload: { data: { attributes: {} } },
+    };
+
+    beforeEach(function () {
+      sinon.stub(usecases, 'reconcileScoOrganizationLearnerAutomatically');
+      usecases.reconcileScoOrganizationLearnerAutomatically.resolves();
+      sinon.stub(scoOrganizationLearnerSerializer, 'serializeIdentity');
+      scoOrganizationLearnerSerializer.serializeIdentity.resolves();
+      sinon.stub(organizationLearnerUserAssociationSerializer, 'serialize');
+      organizationLearnerUserAssociationSerializer.serialize.resolves();
+    });
+
+    it('should return information about deprecation when old route is used', async function () {
+      // when
+      hFake.request = {
+        path: '/api/schooling-registration-user-associations/auto',
+      };
+      const response = await scoOrganizationLearnerController.reconcileScoOrganizationLearnerAutomatically(
+        request,
+        hFake
+      );
+
+      // then
+      expect(response.headers['Deprecation']).to.equal('true');
+      expect(response.headers['Link']).to.equal(
+        '/api/sco-organization-learners/association/auto; rel="successor-version"'
+      );
+    });
+
+    it('should not return information about deprecation when new route is used', async function () {
+      // when
+      hFake.request = {
+        path: '/api/sco-organization-learners/association/auto',
+      };
+      const response = await scoOrganizationLearnerController.reconcileScoOrganizationLearnerAutomatically(
+        request,
+        hFake
+      );
+
+      // then
+      expect(response.headers['Deprecation']).to.not.exist;
+    });
+  });
 });
